rename ConectionRefreshTask to snake_case
