+++ arches_project/core/utils/connection_refresh.py
@@ -1,6 +1,6 @@
 from qgis.core import QgsApplication
 from PyQt5.QtCore import QTimer
-from arches_project.core.arches.ConnectionRefreshTask import ConnectionRefreshTask
+from arches_project.core.arches.connection_refresh_task import ConnectionRefreshTask
 from arches_project.core.views.components.dialog_updates import reset_refresh_btn
